requirements.txt
python-dotenv
pip install fastapi uvicorn python-dotenv
pip install llama-index langchain google-generativeai
pip install pymongo neo4j
pip install streamlit
pymongo
neo4j
certifi
pip install pymupdf langchain-text-splitters